old-tests/submission/Opera/preload/metadata/preload-metadata-buffered-ogg.html

<!doctype html>
<title>HTMLMediaElement.preload:metadata 'buffered' check - ogg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata 'buffered' check - ogg", "HTMLVideoElement.preload:metadata 'buffered' check - ogg", {timeout:7000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
	var tag_name = vars[1];
	var events_actual = [];
	var max_buffer = 60;
	var tolerance = 4;
	var node = document.getElementsByTagName(tag_name)[0];
	startTest();

	function startTest() {
	    var events  = ['loadstart', 'progress', 'suspend', 'abort', 'error', 'loadedmetadata', 'loadeddata', 'canplay', 'canplaythrough'];
	    for (var e in events) {
		node.addEventListener(events[e], handleEvent, false);
	    }
	    node.src = '../resources/preload.ogv?pipe=trickle(100000:d1:r2)&nocache=' + Math.random();
	}

	function handleEvent(e) {
	    if (e.type == 'error' && !node.canPlayType(tag_name + '/$media_type')) {
		t.step(function() {
		    assert_unreached("Cannot play '" + tag_name + "/$media_type'");
		    t.done();
		});
	    }
	    if (e.type == 'suspend') endTest();
	}

	function endTest() {
	    t.step(function() {
		assert_equals(node.buffered.length, 2, "'buffered' should have a length of 2");
		assert_equals(node.buffered.start(0), 0, "The start of first buffered timerange should be 0");
		assert_true(node.buffered.end(0) < max_buffer, "The end of first buffered timerange should be less than " + max_buffer + " seconds");
		assert_true(Math.abs(node.buffered.start(1) - node.duration) < tolerance, "The start of second buffered timerange should be media.duration \u00B1 " + tolerance);
		assert_true(Math.abs(node.buffered.end(1) - node.duration) < tolerance, "The end of second buffered timerange should be media.duration \u00B1 " + tolerance);
		t.done();
	    });
	}
    });
});
</script>
